chore: refactor frost/eddsa/redjubjub folders
