+++ src/lib.rs
@@ -3,8 +3,8 @@ pub mod participants;
 
 pub mod confidential_key_derivation;
 pub mod ecdsa;
-pub mod eddsa;
 pub mod errors;
+pub mod frost;
 
 #[cfg(feature = "test-utils")]
 pub mod test_utils;
